Reflect package/module filename
The way we were compiling there was no reference to the original file in
the compiled code, so when an exception or a logger tried to display the
line it wouldn't know the name of the original file and would just
report <string>.

This patch fixes that and now we'll see the original file location on
exceptions and logging lines.

We still cannot see the contents of the file on exceptions.