Show code on tracebacks and debugging
This patch adds the mechanism to display the source code when debugging
and on tracebacks.

Due to how CPython handles uncaught exceptions we will need to install a
hook to display the code.

This hook is set automatically when using CPython, for example iPython
has no need for our hook since it already has one that uses normal
traceback and linecache.

We can disable automatically setting this hook with --set-except and
--no-except optional arguments.