Reduce memory footprint
We were storing the package structure and all the code in a dictionary,
which meant that once everything was loaded, compiled and running we no
longer needed to use that memory anymore, but we still used it.

This patch changes the way we store the data by using the file as the
container and storing the files inside global strings that that use no
memory in the dictionary.

The tree will still keep the package structure -although now we store it
flat, to have faster access- but we'll only store 3 integers in the
dictionary. One to mark if the element is a package, another to mark the
byte where the source code begins and the last one to mark the end of
the source code.

Then when we want to load a package we just open the file and jump
straight to where the code is and read just the amount of bytes we need
to to compile the package/module.