Generate and use .pyc files
This patch adds the pyc file functionality as common files do.

If the pyc file does not exist, we'll try to create one.

pyc file will be used if it exists, and the magic number stored in the
pyc file matches executing interpreter, and original source code
modification time has not changed from the compiled file.