Fix default package import
When we import the package the `prepare_package` is automatically run to
load our own importer and to load and compile the original `__init__.py`
file instead of the pinliner version.

On the last step of this process when we execute the compiled code to
replace the running one we were incorrectly running it in the current
scope, which made us lose all the attributes of the module.

This means that if our `__init__.py` defined a variable `my_var` and we
did `import a` then we wouldn't be able to see `a.a`.

This patch fixes this by executing the code using the right global
dictionary.