Close #1: Add multiple package support
Add support to work as a bundle where we can have multiple packages in a
single file.

When working as a bundle we'll default to not auto-load any of the
included packages when the bundle is loaded, though this can be changed
using '--default-pkg' optional argument.

To define the output file we now need to use '--outfile' argument, as we
have no way of differentiating when it's the last file we want to
include and when it's the output filename.